dailybuild: Projectファイルの修正 （大半はVS2008版で.pdbファイルの置き場所を変えたことによるもの）
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@7242 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/tests/Graphics/GRSimple/main.cpp
@@ -108,7 +108,7 @@ void __cdecl display(){
 	//-----------------------------------
 	render->SetMaterial(matLine);
 	Vec3f vtx[4] = {Vec3f(0,0,0), Vec3f(10,0,0), Vec3f(0,10,0), Vec3f(0,0,10)};
-	size_t vtxIndex[6] = {0, 1, 0, 2, 0, 3};
+	GLuint vtxIndex[6] = {0, 1, 0, 2, 0, 3};
 	render->SetLineWidth(2.0);
 	render->SetVertexFormat(GRVertexElement::vfP3f);
 	render->DrawIndexed(render->LINES, vtxIndex, vtx, 6);
